Add custom simple middleware to check for user role

+++ routes/api.php
@@ -22,10 +22,10 @@ use App\Http\Controllers\API\UserController;
 Route::post('login', [UserController::class, 'login']);
 Route::post('logout', [UserController::class, 'logout'])->middleware('auth:sanctum');
 
-Route::group(['prefix' => 'users', 'middleware' => 'auth:sanctum'], function () {
-    Route::get('/', [UserController::class, 'index']);
-    Route::post('add', [UserController::class, 'register']);
-    Route::get('edit/{id}', [UserController::class, 'edit']);
-    Route::post('update/{id}', [UserController::class, 'update']);
-    Route::delete('delete/{id}', [UserController::class, 'delete']);
+Route::group(['prefix' => 'users'], function () {
+    Route::get('/', [UserController::class, 'index'])->middleware(['userRole:super-admin']);
+    Route::post('add', [UserController::class, 'register'])->middleware(['userRole:super-admin']);
+    Route::get('edit/{id}', [UserController::class, 'edit'])->middleware(['userRole:super-admin']);
+    Route::post('update/{id}', [UserController::class, 'update'])->middleware(['userRole:super-admin']);
+    Route::delete('delete/{id}', [UserController::class, 'delete'])->middleware(['userRole:super-admin']);
 });
